1、收货地址、竟拍出价、竟拍出价保证金、竟拍活动、竟拍活动提醒、商品关注=》实体Id或关联Id类型、Mapper方法传参类型都调整为Integer; 实体ID调整为@GeneratedValue(strategy = GenerationType.IDENTITY)生成

+++ src/main/java/com/hrtx/web/mapper/AuctionMapper.java
@@ -11,47 +11,47 @@ import java.util.Map;
 
 public interface AuctionMapper extends Mapper<Auction>,BaseMapper<Auction>{
 
-    List<Map> findAuctionSumEPSaleGoodsByNumId(@Param("numId") Long numId);
+    List<Map> findAuctionSumEPSaleGoodsByNumId(@Param("numId") Integer numId);
 
-    List<Map> findAuctionSumEPSaleGoodsByNumIdAndGId(@Param("numId") Long numId,@Param("gId") Long gId);
+    List<Map> findAuctionSumEPSaleGoodsByNumIdAndGId(@Param("numId") Integer numId,@Param("gId") Integer gId);
 
-    List<Map> findAuctionSumEPSaleGoodsByGId(@Param("gId") Long gId);
+    List<Map> findAuctionSumEPSaleGoodsByGId(@Param("gId") Integer gId);
 
-    List<Map> findAuctionListByNumId(@Param("numId") Long numId);
+    List<Map> findAuctionListByNumId(@Param("numId") Integer numId);
 
-    List<Map> findAuctionListByNumIdAndGId(@Param("numId") Long numId,@Param("gId") Long gId);
+    List<Map> findAuctionListByNumIdAndGId(@Param("numId") Integer numId,@Param("gId") Integer gId);
 
-    List<Map> findAuctionListByGId(@Param("gId") Long gId);
+    List<Map> findAuctionListByGId(@Param("gId") Integer gId);
 
-    List<Map> findAuctionListByNumIdAndGId2(@Param("numId") Long numId,@Param("gId") Long gId);
+    List<Map> findAuctionListByNumIdAndGId2(@Param("numId") Integer numId,@Param("gId") Integer gId);
 
-    List<Map> findAuctionListByGId2(@Param("gId") Long gId);
+    List<Map> findAuctionListByGId2(@Param("gId") Integer gId);
 
-    List<Map> findAuctionListByNumIdAndConsumerIdAndGId(@Param("numId") Long numId,@Param("consumerId") Long consumerId,@Param("gId") Long gId);
+    List<Map> findAuctionListByNumIdAndConsumerIdAndGId(@Param("numId") Integer numId,@Param("consumerId") Integer consumerId,@Param("gId") Integer gId);
 
-    List<Map> findAuctionListByConsumerIdAndGId(@Param("consumerId") Long consumerId,@Param("gId") Long gId);
+    List<Map> findAuctionListByConsumerIdAndGId(@Param("consumerId") Integer consumerId,@Param("gId") Integer gId);
 
     List<Map> findAuctionOrderListByConsumerId(Auction auction);
 
-    List<Map>  findAuctionListByOrderId(@Param("orderId") Long orderId);
+    List<Map>  findAuctionListByOrderId(@Param("orderId") Integer orderId);
 
-    Page<Object> queryPageNumList(Num num, @Param("consumerId") Long consumerId, @Param("status") String status);
+    Page<Object> queryPageNumList(Num num, @Param("consumerId") Integer consumerId, @Param("status") String status);
 
-    List<Map>  findAuctionListByNumIdAndPrice(@Param("numId") Long numId,@Param("price") double price);
+    List<Map>  findAuctionListByNumIdAndPrice(@Param("numId") Integer numId,@Param("price") double price);
 
-    List<Map>  findAuctionListByNumIdAndPrice2(@Param("numId") Long numId,@Param("price") double price);
+    List<Map>  findAuctionListByNumIdAndPrice2(@Param("numId") Integer numId,@Param("price") double price);
 
-    List<Map> findAuctionListDepositByNumId(@Param("numId") Long numId);
+    List<Map> findAuctionListDepositByNumId(@Param("numId") Integer numId);
 
-    List<Map> findAuctionListDepositByNumId2(@Param("numId") Long numId);
+    List<Map> findAuctionListDepositByNumId2(@Param("numId") Integer numId);
 
-    List<Map> findAuctionListDepositByNumIdAndGId(@Param("numId") Long numId,@Param("gId") Long gId);
+    List<Map> findAuctionListDepositByNumIdAndGId(@Param("numId") Integer numId,@Param("gId") Integer gId);
 
-    List<Map> findAuctionListDepositByNumIdAndGId2(@Param("numId") Long numId,@Param("gId") Long gId);
+    List<Map> findAuctionListDepositByNumIdAndGId2(@Param("numId") Integer numId,@Param("gId") Integer gId);
 
-    List<Map> findAuctionGoodsByNumId(@Param("numId") Long numId);
+    List<Map> findAuctionGoodsByNumId(@Param("numId") Integer numId);
 
-    List<Map> findCustomersByNumIdAndGId(@Param("numId") Long numId,@Param("gId") Long gId);
+    List<Map> findCustomersByNumIdAndGId(@Param("numId") Integer numId,@Param("gId") Integer gId);
 
     List<Map>  findAuctionByNumIdAndStatus(Auction auction);
 
@@ -59,7 +59,7 @@ public interface AuctionMapper extends Mapper<Auction>,BaseMapper<Auction>{
 
     void insertBatch(@Param("auctionList") List<Auction> list);
 
-   // void auctionEditStatusById(@Param("status") int  status,@Param("id") Long id);
+   // void auctionEditStatusById(@Param("status") int  status,@Param("id") Integer id);
 
     void auctionEditStatusById(Auction auction);
 
@@ -73,5 +73,5 @@ public interface AuctionMapper extends Mapper<Auction>,BaseMapper<Auction>{
 
     void auctionEditOrderIDByGId(Auction auction);
 
-    List<Map> freezeOneNum(@Param("numId") Long numId);
+    List<Map> freezeOneNum(@Param("numId") Integer numId);
 }
